requirements.txt
aiohttp
tortoise-orm[asyncpg]
pyjwt[crypto]